Fix paths in "include" for moved repo implementation files
