+++ libdnf/repo/config_repo.cpp
@@ -17,7 +17,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#include "libdnf/rpm/config_repo.hpp"
+#include "libdnf/repo/config_repo.hpp"
 
 #include "../conf/config_utils.hpp"
