+++ libdnf/repo/repo_sack.cpp
@@ -17,7 +17,7 @@ You should have received a copy of the GNU Lesser General Public License
 along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#include "libdnf/rpm/repo_sack.hpp"
+#include "libdnf/repo/repo_sack.hpp"
 
 #include "repo_impl.hpp"
