+++ libdnf/rpm/solv/solv_private.hpp
@@ -22,7 +22,7 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 
 
 #include "../package_sack_impl.hpp"
-#include "../repo_impl.hpp"
+#include "../../repo/repo_impl.hpp"
 
 namespace libdnf::rpm::solv {
