+++ libdnf/rpm/transaction.cpp
@@ -24,7 +24,7 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 #include "../libdnf/utils/bgettext/bgettext-lib.h"
 #include "package_sack_impl.hpp"
 #include "package_set_impl.hpp"
-#include "repo_impl.hpp"
+#include "../repo/repo_impl.hpp"
 #include "solv/id_queue.hpp"
 
 #include <fcntl.h>
