+++ microdnf/commands/install/install.cpp
@@ -28,7 +28,6 @@ along with microdnf.  If not, see <https://www.gnu.org/licenses/>.
 #include <libdnf/rpm/package.hpp>
 #include <libdnf/rpm/package_query.hpp>
 #include <libdnf/rpm/package_set.hpp>
-#include <libdnf/rpm/repo.hpp>
 #include <libdnf/rpm/transaction.hpp>
 
 #include <filesystem>
