+++ test/libdnf/advisory/test_advisory_query.hpp
@@ -22,7 +22,7 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 #define TEST_LIBDNF_ADVISORY_ADVISORY_QUERY_HPP
 
 
-#include "../rpm/repo_fixture.hpp"
+#include "../repo_fixture.hpp"
 
 #include "libdnf/advisory/advisory_sack.hpp"
