+++ test/libdnf/repo/test_repo.cpp
@@ -22,7 +22,7 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 #include "libdnf/base/base.hpp"
 #include "libdnf/logger/stream_logger.hpp"
 #include "libdnf/rpm/package_sack.hpp"
-#include "libdnf/rpm/repo_sack.hpp"
+#include "libdnf/repo/repo_sack.hpp"
 
 #include <filesystem>
 #include <fstream>
