+++ test/libdnf/repo/test_repo.hpp
@@ -20,7 +20,7 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 #ifndef LIBDNF_TEST_REPO_REPO_HPP
 #define LIBDNF_TEST_REPO_REPO_HPP
 
-#include "../../testcase_fixture.hpp"
+#include "../testcase_fixture.hpp"
 
 #include "libdnf/utils/temp.hpp"
