+++ test/libdnf/rpm/test_package.hpp
@@ -22,7 +22,7 @@ along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 #define TEST_LIBDNF_RPM_PACKAGE_HPP
 
 
-#include "repo_fixture.hpp"
+#include "../repo_fixture.hpp"
 
 #include "libdnf/rpm/package.hpp"
